* fix bug with Content-Length estimation when returned body is not String
There was a bug that for body which wasn't instance of String Content-Length
was wrong estimated so for ["Hello"] it was 1 instead 5.

Rack spec says only that returned body from app.call is object that
respond to "each" and calling "each" should yield String object.
If we want handle all cases we should compute Content-Length calling
"each" and adding "length" of each part.